third_party/WebKit/LayoutTests/svg/animations/animate-fill-freeze-with-repeatDur-expected.txt

SVG 1.1 dynamic animation tests

Test for animation freeze when repeatDur is not a multiple of dur

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS rect1.x.animVal.value is 150
PASS rect1.x.baseVal.value is 0
PASS successfullyParsed is true

TEST COMPLETE
